Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    float x;
-    scanf("%f",&x);
-    printf("You entered: %.4f",x);
+    double x;
+    scanf("%lf",&x);
+    printf("You entered: %.4lf",x);
     return 0;
     
 }
